Migrated MessageDelete to the BotEvent structure
